refactor: Change profile directory
