+++ app/src/main/res/layout/activity_mentee_to_mentor.xml
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+    <data>
+
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <androidx.appcompat.widget.Toolbar
+            android:id="@+id/toolbar_title"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/size_toolbar_height"
+            app:contentInsetStart="@dimen/margin_component_inset_20dp"
+            app:contentInsetEnd="@dimen/margin_component_inset_20dp"
+            android:background="@color/white"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent">
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+                <ImageButton
+                    android:id="@+id/ib_back"
+                    android:layout_width="@dimen/size_icon_button_24dp"
+                    android:layout_height="@dimen/size_icon_button_24dp"
+                    android:src="@drawable/icon_back"
+                    android:background="@color/transparent"
+                    android:scaleType="fitXY"
+
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintBottom_toBottomOf="parent" />
+                <TextView
+                    android:id="@+id/tv_title"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    style="@style/CustomTextViewMidBlack"
+                    android:text="@string/tv_account_withdrawal"
+                    android:textSize="@dimen/size_tv_title_15dp"
+                    android:fontFamily="@font/noto_sans_kr_medium"
+                    android:background="@color/transparent"
+                    android:gravity="center"
+
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintBottom_toBottomOf="parent" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+        </androidx.appcompat.widget.Toolbar>
+
+        <TextView
+            android:id="@+id/tv_encourage_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/tv_account_withdrawal_encourage_title"
+            android:textSize="@dimen/size_18dp"
+            android:textColor="@color/mid_black"
+            android:fontFamily="@font/noto_sans_kr_bold"
+            android:includeFontPadding="false"
+            android:gravity="center"
+            android:layout_marginTop="@dimen/size_20dp"
+            app:layout_constraintTop_toBottomOf="@id/toolbar_title"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent" />
+        <ImageView
+            android:id="@+id/iv_graphic"
+            android:layout_width="@dimen/size_100dp"
+            android:layout_height="@dimen/size_110dp"
+            android:src="@drawable/ic_graphic_withdrawal"
+            android:layout_marginTop="@dimen/size_73dp"
+
+            app:layout_constraintTop_toBottomOf="@id/tv_encourage_title"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent" />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/tv_account_withdrawal_encourage_content"
+            android:textSize="@dimen/size_12dp"
+            android:textColor="@color/mid_black"
+            android:fontFamily="@font/noto_sans_kr_medium"
+            android:includeFontPadding="false"
+            android:gravity="center"
+            android:layout_marginTop="@dimen/size_50dp"
+            app:layout_constraintTop_toBottomOf="@id/iv_graphic"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent" />
+        
+        
+        <Button
+            android:id="@+id/btn_next"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/size_50dp"
+            android:background="@drawable/selector_btn_bg_purple_active_gray_rectangle"
+            android:text="@string/btn_next"
+            android:textColor="@color/white"
+            android:textSize="@dimen/size_15dp"
+            android:includeFontPadding="false"
+            android:fontFamily="@font/noto_sans_kr_medium"
+            app:layout_constraintBottom_toBottomOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
